% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustbetareg.R
\name{degb}
\alias{degb}
\alias{pegb}
\alias{qegb}
\alias{regb}
\title{The EGB of the second type}
\usage{
degb(y_star, mu, phi, log = FALSE)

pegb(q, mu, phi)

qegb(p, mu, phi)

regb(n, mu, phi)
}
\arguments{
\item{y_star, q}{vector of quantiles.}

\item{mu}{mu parameter (\eqn{\mu\in(0,1)}).}

\item{phi}{phi parameter (\eqn{\phi>0}).}

\item{log}{a logical value. If TRUE return the log of density function.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{degb} gives the density, \code{pegb} gives the distribution function, \code{qegb} gives the quantile function, and \code{regb} generates random sample of EGB variables.
}
\description{
Density and random generation for the exponential generalized beta (EGB) of the second type.
}
\details{
The EGB density function is \eqn{f_{\theta}(y^{\star};\mu,\phi)=B^{-1}(\mu\phi,(1-\mu)\phi) \exp\{-y^{\star}(1-\mu)\phi\}/ (1+\exp\{-y^{\star}\})^{\phi}}, with \eqn{\mu\in(0,1),\phi>0} and \eqn{y^{\star}\in R} where \eqn{E(y^{\star})=\psi(\mu\phi)-\psi((1-\mu)\phi)}, \eqn{Var(y^{\star})=\psi'(\mu\phi)+\psi'((1-\mu)\phi)} where \eqn{\psi} is the digamma function. See Kerman and McDonald (2015) for more details.
}
\references{
\href{https://doi.org/10.1080/03610926.2013.844255}{Kerman, S. McDonald, J. B. Skewness-kurtosis bounds for EGB1, EGB2, and special cases. Communications in Statistics - Theory and Methods, 44:3857-3864 (2015).}
}
