% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{degbeta}
\alias{degbeta}
\title{The EGB of the second type}
\usage{
degbeta(y_star, mu, phi, log)
}
\arguments{
\item{y_star}{Logit transformation of original data}

\item{mu}{mu parameter}

\item{phi}{phi parameter}

\item{log}{LOGICAL parameter. If TRUE return the log of density function}
}
\value{
Return value of density function.
}
\description{
Exponential Generalized Beta (EGB) of the second type density function.
}
\details{
For more details see
}
