% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firm.R
\docType{data}
\name{Firm}
\alias{Firm}
\title{Firm Cost}
\format{
A data frame with 73 rows and 7 variables:
\describe{
  \item{FIRMCOST}{total property and casualty premiums and uninsured losses
     as a percentage of total assets.}
  \item{ASSUME}{per occurrence retention amount as a percentage of total
     assets.}
  \item{CAP}{indicates that the firm owns a captive insurance company; 1 if
     the firm uses a captive, 0 otherwise.}
  \item{SIZELOG}{logarithm of total assets.}
  \item{INDCOST}{a measure of the firm's industry risk.}
  \item{CENTRAL}{a measure of the importance of the local managers in
     choosing the amount of risk to be retained.}
  \item{SOPH}{a measure of the degree of importance in using analytical tools.}
}
}
\source{
\url{https://instruction.bus.wisc.edu/jfrees/jfreesbooks/Regression\%20Modeling/BookWebDec2010/CSVData/RiskSurvey.csv}
}
\usage{
data("Firm", package = "robustbetareg")
}
\description{
A dataset on risk management practices of 73 firms.
}
\details{
The dataset was introduced and analyzed by Schmit and Roth (1990) and
    is available in the personal web page of Professor E. Frees (Wisconsin School
     of Business Research). The response variable is \code{FIRMCOST}, smaller
     values of firm cost are attributed to firms that have a good risk management
     performance.
}
\references{
Schmit, J.T. and Roth, K. (1990). Cost effectiveness of risk management
    practices. \emph{Journal of Risk and Insurance}. 57:455-470.
}
\keyword{datasets}
