% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{h1}
\alias{h1}
\title{Inv-Logit Link function}
\usage{
h1(x, B)
}
\arguments{
\item{x}{Design matrix}

\item{B}{Beta parameter vector}
}
\value{
Return mu parameter for each observation
}
\description{
Inverse logit link function for parameter mu
}
\details{
This function... - RobustBetaReg
}
