% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{h2}
\alias{h2}
\title{Inv-log Link function}
\usage{
h2(z, G)
}
\arguments{
\item{z}{Design matrix}

\item{G}{Gamma parameter vector}
}
\value{
Return mu parameter for each observation
}
\description{
Inverse log link function for parameter phi
}
\details{
This function... - RobustBetaReg
}
