% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{Initial.points}
\alias{Initial.points}
\title{Robust Initial Points}
\usage{
Initial.points(y, X, Z)
}
\arguments{
\item{y}{...}

\item{X}{...}

\item{Z}{...}
}
\value{
Return value ...
}
\description{
Robust Initial Points
}
