% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-LMDPDE.R
\name{LMDPDE_Wald_Type}
\alias{LMDPDE_Wald_Type}
\title{Robust Wald type Test - LMDPDE}
\usage{
LMDPDE_Wald_Type(M, m, y, x, z, start_theta, alpha, tolerance)
}
\arguments{
\item{M}{teste}

\item{m}{teste}

\item{y}{The numeric response vector (with values in (0,1)).}

\item{x}{The numeric regressor matrix for mean model.}

\item{z}{The numeric regressor matrix for precision model, defaulting to an intercept only.}

\item{start_theta}{A numeric vector with an initial guess of the root of estimation equation.}

\item{alpha}{The tuning with values (0,1), for robust estimation. When alpha is equal zero is equivalent of MLE.}

\item{tolerance}{The function value tolerance.}
}
\value{
Return a list of components:
\itemize{
\item coefficients - a vector with elements "mean" and "precision" containing the coefficients from the respective models.
\item vcov - covariance matrix of all parameters in the model.
}
}
\description{
Wald type test statistic based on LMDPDE
}
\details{
For more details see:
}
