% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-LSMLE.R
\name{LSMLE.Beta.Reg}
\alias{LSMLE.Beta.Reg}
\title{Robust Estimation - LSMLE}
\usage{
LSMLE.Beta.Reg(
  y,
  x,
  z,
  alpha,
  alpha.optimal,
  start_theta,
  L,
  M,
  tolerance,
  maxit
)
}
\arguments{
\item{y}{The numeric response vector (with values in (0,1)).}

\item{x}{The numeric regressor matrix for mean model.}

\item{z}{The numeric regressor matrix for precision model, defaulting to an intercept only.}

\item{alpha}{The tuning with values (0,1), for robust estimation. When alpha is equal zero is equivalent of MLE.}

\item{alpha.optimal}{A logical value. If TRUE the tuning parameter should be selected automatic.}

\item{start_theta}{A numeric vector with an initial guess of the root of estimation equation.}

\item{L}{A parameter of auto selecting algorithm of tuning parameter (default L=0.02).}

\item{M}{A integer parameter value of auto selecting algorithm of tuning parameter (default M=3).}

\item{tolerance}{The function value tolerance.}

\item{maxit}{The maximum number of iterations used by the algorithm.}
}
\value{
Return a list of components:
\itemize{
\item coefficients - the vector with elements "mean" and "precision" containing the coefficients from the respective models,
\item vcov - the covariance matrix of all parameters in the model,
\item mu_hat - the vector of fitted means,
\item phi_hat - the vector of predicted precision,
\item weight - the weights used,
\item Tuning - the employed tuning parameter,
\item Res.Beta - a vector of standardized weighted residual 2,
\item start - the starting values for the parameters passed to the Newton-Raphson algorithm,
\item std.error - the standard error vector of coefficients,
\item Tab - the summary table of overall result.
}
}
\description{
Fit robust beta regression models for rates and proportions via LSMLE.
}
\details{
For more details see:...
}
