% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LSMLE.R
\name{LSMLE_Cov_Matrix}
\alias{LSMLE_Cov_Matrix}
\title{Sandwich Matrix - LSMLE}
\usage{
LSMLE_Cov_Matrix(mu, phi, X, Z, alpha)
}
\arguments{
\item{mu}{The numeric mean parameter.}

\item{phi}{The numeric precision parameter.}

\item{X}{The numeric regressor matrix for mean model.}

\item{Z}{The numeric regressor matrix for precision model, defaulting to an intercept only.}

\item{alpha}{The tuning with values (0,1), for robust estimation. When alpha is equal to zero, it is equivalent to MLE.}
}
\value{
Return a tuning given by data driven algorithm selection
}
\description{
Covariance matrix of LSMLE.
}
\details{
For more details see:
}
