% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{Newton.Raphson}
\alias{Newton.Raphson}
\title{Newton-Raphson}
\usage{
Newton.Raphson(x, M, tol, details, FUN = f, ...)
}
\arguments{
\item{M}{Max number of iteration}

\item{tol}{Tolerance}

\item{details}{A logical indicating if the each iteration should be return.}

\item{FUN}{A function of x returning a vector of function values with the same length as the vector x.}

\item{p0}{Initial point}
}
\value{
Return solution of a system of nonlinear equations.
}
\description{
Newton-Raphson algorithm. The function solves a system of nonlinear equations.
}
