% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LSMLE.R
\name{Opt.Tuning.LSMLE}
\alias{Opt.Tuning.LSMLE}
\title{Auto Selecting tuning parameter algorithm}
\usage{
Opt.Tuning.LSMLE(y, x, z, tolerance, L, M)
}
\arguments{
\item{y}{The numeric response vector (with values in (0,1)).}

\item{x}{The numeric regressor matrix for mean model.}

\item{z}{The numeric regressor matrix for precision model, defaulting to an intercept only.}

\item{tolerance}{The function value tolerance.}

\item{L}{A parameter of auto selecting algorithm of tuning parameter (default L=0.02).}

\item{M}{A integer parameter value of auto selecting algorithm of tuning parameter (default M=3).}
}
\value{
Return a tuning given by data driven algorithm selection
}
\description{
The algorithm of auto-selecting tuning.
}
\details{
For more details see:
}
