% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show-LMDPDE.R, R/Show-LSMLE.R
\name{plot.LMDPDE}
\alias{plot.LMDPDE}
\alias{plot.LSMLE}
\title{Interactive plots for diagnostic of robust betareg models}
\usage{
plot.LMDPDE(object, ask = TRUE, ...)

plot.LSMLE(object, ask = TRUE, ...)
}
\arguments{
\item{object}{fitted model object of class  "LSMLE" or "LMDPDE".}

\item{ask}{logical. If TRUE the user is asked before each plot.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Several types of standard diagnostic plots can be produced interactively, involving various kinds of residuals, influence measures, weights etc.
}
\examples{
fit=robustbetareg(I(food/income)~income+persons|1,data=FoodExpenditure,alpha=0.08)
plot(fit)

}
