% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{plotenvelope}
\alias{plotenvelope}
\title{Simulated Envelope of Residuals}
\usage{
plotenvelope(robustbetareg.obj, n.sim, ylim, index, control)
}
\arguments{
\item{robustbetareg.obj}{Object of robust beta regression (see \code{\link[RobustBetaReg:robustbetareg]{robustbetareg}}).}

\item{n.sim}{the number of simulation sample}

\item{ylim}{the y limits of the plot}

\item{index}{index of points to draw}

\item{control}{a list of control arguments specified via \code{\link[RobustBetaReg:robustbetareg.control]{robustbetareg.control}}.}
}
\value{
Return a simulated envelope graphic.
}
\description{
Plot a simulated envelope of beta residuals, from LMDPDE and LSMLE.
}
\examples{
rbr.obj=robustbetareg(I(food/income)~income+persons|1,data=FoodExpenditure,alpha=0.08)
plotenvelope(rbr.obj,n.sim=100)

}
