% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(
  object,
  newdata = NULL,
  type = c("response", "link", "precision", "variance", "quantile"),
  at = 0.5
)
}
\arguments{
\item{object}{Fitted model object of class "LSMLE" or "LMDPDE" (see \code{\link[RobustBetaReg:robustbetareg]{robustbetareg}}).}

\item{n.sim}{the number of simulation sample. Deafault n.sim=50.}

\item{conf}{the confidence level of the envelopes required. The default is to find 95 confidence envelopes.}

\item{control}{a list of control arguments specified via \code{\link[RobustBetaReg:robustbetareg.control]{robustbetareg.control}}.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
Return a simulated envelope graphic.
}
\description{
Plot a simulated envelope of beta residuals, from LSMLE and LMDPDE objects.
}
\examples{
fit=robustbetareg(I(food/income)~income+persons|1,data=FoodExpenditure,alpha=0.08)
plotenvelope(fit,n.sim=100)

}
