% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LMDPDE.R
\name{Psi_Beta}
\alias{Psi_Beta}
\title{Score Function LMDPDE - Beta}
\usage{
Psi_Beta(Beta, Gamma, y, X, Z, alpha)
}
\arguments{
\item{Beta}{...}

\item{Gamma}{...}

\item{y}{...}

\item{X}{...}

\item{Z}{...}

\item{alpha}{...}
}
\value{
Return value ...
}
\description{
Modified Score Function LMDPDE - Beta
}
\details{
For more details see:
}
