% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LSMLE.R
\name{Psi_Beta_LSMLE}
\alias{Psi_Beta_LSMLE}
\title{Score Function LSMLE - Beta}
\usage{
Psi_Beta_LSMLE(Beta, Gama, y, X, Z, alpha)
}
\arguments{
\item{Beta}{...}

\item{y}{...}

\item{X}{...}

\item{Z}{...}

\item{alpha}{...}

\item{Gamma}{...}
}
\value{
Return value ...
}
\description{
Modified Score Function LSMLE - Beta
}
\details{
For more details see:
}
