% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LMDPDE.R
\name{Psi_LMDPDE}
\alias{Psi_LMDPDE}
\title{Modified Score Vector - LMDPDE}
\usage{
Psi_LMDPDE(Theta, y, X, Z, alpha)
}
\arguments{
\item{Theta}{...}

\item{y}{...}

\item{X}{...}

\item{Z}{...}

\item{alpha}{...}
}
\value{
Return value ...
}
\description{
Modified Score Function LMDPDE - Gamma
}
\details{
For more details see:
}
