% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LSMLE.R
\name{Psi_LSMLE}
\alias{Psi_LSMLE}
\title{Modified Score Vector - LSMLE}
\usage{
Psi_LSMLE(Theta, y, X, Z, alpha)
}
\arguments{
\item{Theta}{...}

\item{y}{...}

\item{X}{...}

\item{Z}{...}

\item{alpha}{...}
}
\value{
Return value ...
}
\description{
Modified Score Function LSMLE - Gamma
}
\details{
For more details see:
}
