% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{regbeta}
\alias{regbeta}
\title{The EGB of the second type}
\usage{
regbeta(n, mu, phi)
}
\arguments{
\item{n}{number of observations}

\item{mu}{mu parameter}

\item{phi}{phi parameter}
}
\value{
Return value of density function.
}
\description{
Random generation for Exponential Generalized Beta (EGB) of the second type.
}
\details{
For more details see
}
