% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{Residual_Beta}
\alias{Residual_Beta}
\title{Residual Beta}
\usage{
Residual_Beta(mu_hat, phi_hat, y, X, Z)
}
\arguments{
\item{mu_hat}{...}

\item{phi_hat}{...}

\item{y}{...}

\item{X}{...}

\item{Z}{...}
}
\value{
Return value ...
}
\description{
Residual Type Standardized Weighted Residual 2.
}
\details{
For more details see: On beta regression residuals: Patrícia L. Espinheira a , Silvia L.P. Ferrari and Francisco Cribari-Neto; Journal of Applied Statistics Vol. 35, No. 4, April 2008
}
