% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{Residual_Gamma}
\alias{Residual_Gamma}
\title{Residual Gamma}
\usage{
Residual_Gamma(mu_hat, phi_hat, y, X, Z)
}
\arguments{
\item{mu_hat}{...}

\item{phi_hat}{...}

\item{y}{...}

\item{X}{...}

\item{Z}{...}
}
\value{
Return value ...
}
\description{
Residual Type Gamma
}
\details{
For more details see: Espinheira, P. L., Santos, E. G., and Cribari Neto, F. (2017). On nonlinear beta regression residuals. Biometrical Journal, 59:445–461.
}
