% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-LMDPDE.R
\name{Robst.LMDPDE.Beta.Reg}
\alias{Robst.LMDPDE.Beta.Reg}
\title{Robust Point Estimation - LMDPDE}
\usage{
Robst.LMDPDE.Beta.Reg(y, x, z, start_theta, alpha, tolerance, maxit)
}
\arguments{
\item{y}{...}

\item{x}{...}

\item{z}{...}

\item{alpha}{...}

\item{Theta}{...}
}
\value{
Return value ...
}
\description{
Robust point estimation by estimating equation via LSMLE
}
\details{
For more details see:
}
