% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{robustbetareg.control}
\alias{robustbetareg.control}
\title{Control Parameter for Robust Beta Regression}
\usage{
robustbetareg.control(tolerance = 0.001, maxit = 250, L = 0.02, M = 3, ...)
}
\arguments{
\item{tolerance}{numeric tolerance for convergence.}

\item{maxit}{integer specifying the maxit argument of iterations used by the Newton-Raphson algorithm.}

\item{L}{A parameter of auto selecting algorithm of tuning parameter (default L=0.02).}

\item{M}{A integer parameter value of auto selecting algorithm of tuning parameter (default M=3).}
}
\value{
A list with the arguments specified.
}
\description{
Various parameters that control fitting of robust beta regression models using robustbetareg
}
\details{
For more details see:...
}
