% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{robustbetareg}
\alias{robustbetareg}
\title{Robust Beta Regression}
\usage{
robustbetareg(
  formula,
  data,
  alpha,
  type = c("LMDPDE", "LSMLE"),
  alpha.optimal,
  start,
  control = robustbetareg.control(...),
  ...
)
}
\arguments{
\item{formula}{symbolic description of the model (of type y ~ x or y ~ x | z).}

\item{data}{arguments controlling formula.}

\item{alpha}{the tuning with values (0,1), for robust estimation. When alpha is equal zero is equivalent of MLE.}

\item{type}{character specification of the type of estimator. Currently, LMDPDE (default) and LSMLE.}

\item{alpha.optimal}{a logical value. If TRUE the tuning parameter should be selected automatic.}

\item{start}{a numeric vector with an initial guess of the root of estimation equation.}

\item{control}{a list of control arguments specified via \code{\link[RobustBetaReg:robustbetareg.control]{robustbetareg.control}}.}
}
\value{
Return a list of components:
\itemize{
\item coefficients - the vector with elements "mean" and "precision" containing the coefficients from the respective models
\item vcov - the covariance matrix of all parameters in the model
\item mu_hat - the vector of fitted means
\item phi_hat - the vector of predicted precision
\item weights - the weights used
\item Tuning - the employed tuning parameter
\item Res.Beta - a vector of standardized weighted residual 2
\item start - the starting values for the parameters passed to the Newton-Raphson algorithm
\item std.error - the standard error vector of coefficients
\item Tab - the summary table of overall result.
}
}
\description{
Fit robust beta regression models for rates and proportionsvia LMDPDE and LSMLE using a parametrization with mean (depending through a link function on the covariates) and precision parameter (called phi).
}
\details{
For more details see:...
}
