% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{SaddlepointTest}
\alias{SaddlepointTest}
\title{Robust Saddlepoint Test}
\usage{
SaddlepointTest(object, FUN, ..., thrd)
}
\arguments{
\item{object}{fitted model object of class "LSMLE" or "LMDPDE"}

\item{FUN}{the function representing the null hypothesis to be tested}

\item{...}{further arguments to be passed}

\item{thrd}{number (integer) of threads to speed up the process. If missing the value is autodetected by the available number of multi-core processor}
}
\description{
Saddlepoint tests for both simple and composite hypothesis for independent but non-homogeneous observations, based on LSMLE and LMDPDE.
}
\examples{
fit=robustbetareg(I(food/income)~income+persons|1,data=FoodExpenditure,alpha=0.08)
h0=function(theta,B){c(theta[1],B)}#Hiphothesis to be tested
SaddlepointTest(fit,h0,B=c(0,0))#H0: income=persons=0

}
\references{
\href{https://www.sciencedirect.com/science/article/pii/S0047259X09001183}{Lo, S. N., Ronchetti, E. Robust and accurate inference for generalized linear models. Journal of Multivariate Analysis, 100, 2126–2136 (2009)}
}
