% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{set.link}
\alias{set.link}
\title{Create a Link for beta regression model}
\usage{
set.link(link.mu = "logit", link.phi = "log")
}
\arguments{
\item{link.mu}{character; one of "logit"(default), "probit", "cauchit", "cloglog" and loglog}

\item{link.phi}{character; one of "log"(default) "indentity" and "sqrt"}
}
\value{
A structure with link function, inverse link function, derivative deta/dmu and d2eta/dmu2, for both models: mean and precision.
}
\description{
This function provides several link functions for beta regression models
}
\details{
For more details see:...
}
\examples{
links=set.link(link.mu="cauchit",link.phi="sqrt")
attributes(links)

}
