% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Teste.R
\name{soma}
\alias{soma}
\title{Teste funcao soma}
\usage{
soma(x1, x2)
}
\arguments{
\item{x1}{eh um numero}

\item{x2}{eh um numero}
}
\value{
retorna a soma
}
\description{
Essa funcao soma dois numeros
}
\details{
Essa funcao eh apenas para teste do desenvolvimento do pacote R - RobustBetaReg
}
