% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{SQV}
\alias{SQV}
\title{SQV}
\usage{
SQV(zq, n, p)
}
\arguments{
\item{zq}{Standardized parameter vector}

\item{n}{Sample size}

\item{p}{Parameter vector length}
}
\value{
Return SQV vector.
}
\description{
Standardized Quadratic Variations
}
\details{
This function... - RobustBetaReg
}
