% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{star.obs}
\alias{star.obs}
\title{Teste ***}
\usage{
star.obs(p.valor)
}
\arguments{
\item{p.value}{}
}
\value{
Return number stars ***.
}
\description{
star observation
}
\details{
For more details see
}
