% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show-Result.R
\name{summary.robustbetareg}
\alias{summary.robustbetareg}
\alias{coef.robustbetareg}
\title{Methods for robustbetareg Objects}
\usage{
\method{summary}{robustbetareg}(object, type = "sweighted2", ...)

\method{coef}{robustbetareg}(object, model = c("full", "mean", "precision"), ...)
}
\arguments{
\item{object}{fitted model of class \code{robustbetareg}.}

\item{type}{character specifying type of residuals to be included in the summary output, see \code{\link[robustbetareg:residuals]{residuals.robustbetareg}}.}

\item{...}{currently not used.}

\item{model}{character specifying for which component of the model coefficients/covariance should be extracted.}
}
\description{
Methods for extracting information from fitted robust beta regression model objects of class "\code{robustbetareg}"
}
\details{
A set of standard extractor functions for fitted model objects is available for objects of class "\code{robustbetareg}",
including methods to the generic functions print and summary which print the estimated coefficients along with some further information.
}
\examples{
\dontrun{data("HIC", package="robustbetareg")
fit=robustbetareg(Percent_HIC~Urbanization+GDP_percapita|1,data=HIC,alpha=0.06)
summary(fit)
coef(fit)
}

}
\references{
\href{https://www.tandfonline.com/doi/abs/10.1080/02664760701834931}{Espinheira, P.L., Ferrari, S.L.P., and Cribari-Neto, F. (2008). On Beta Regression Residuals. Journal of Applied Statistics, 35(4), 407–419.}

\href{https://www.tandfonline.com/doi/abs/10.1080/02664760701834931}{Espinheira, P.L., Ferrari, S.L.P., and Cribari-Neto, F. (2008). On Beta Regression Residuals. Journal of Applied Statistics, 35(4), 407–419.}

\href{https://onlinelibrary.wiley.com/doi/abs/10.1002/bimj.201600136}{Espinheira, P.L., Santos, E.G.and Cribari-Neto, F. (2017). On nonlinear beta regression residuals. Biometrical Journal, 59(3), 445-461.}

\href{https://doi.org/10.1177/0962280217738142}{Ghosh, A. Robust inference under the beta regression model with application to health care studies. Statistical Methods in Medical Research, 28:271-888 (2019).}

\href{https://doi.org/10.48550/arXiv.2209.11315}{Maluf, Y. S., Ferrari, S. L., & Queiroz, F. F. (2022). Robust beta regression through the logit transformation. arXiv}

\href{https://doi.org/10.1007/s00362-022-01320-0}{Ribeiro, K. A. T. Ferrari, S. L. P. Robust estimation in beta regression via maximum Lq-likelihood. Statistical Papers (2022).}
}
\seealso{
\code{\link[robustbetareg:robustbetareg]{robustbetareg}}
}
