% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux-Functions.R
\name{WaldTypeTest}
\alias{WaldTypeTest}
\title{Robust Wald-type Tests}
\usage{
WaldTypeTest(object, g, ...)
}
\arguments{
\item{object}{Object of robust beta regression estimator}

\item{g}{A funtion representing the null hypothesis to be tested}

\item{...}{Further arguments to be passed}
}
\description{
Wald-type tests for both simple and composite hypothesis for independent but non-homogeneous observations, based on LMDPDE and LSMLE.
}
\examples{
rbr.obj=robustbetareg(I(food/income)~income+persons|1,data=FoodExpenditure,alpha=0.08)
g=function(theta,B){theta[2:3]-B}#Hiphothesis to be tested
WaldTypeTest(rbr.obj,g,B=c(0,0))#Testing income=persons=0

}
\references{
\url{https://link.springer.com/article/10.1007\%2Fs00184-018-0653-4}
}
