% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{degbeta}
\alias{degbeta}
\alias{regbeta}
\title{The EGB of the second type}
\usage{
degbeta(y_star, mu, phi, log = FALSE)

regbeta(n, mu, phi)
}
\arguments{
\item{y_star}{logit transformation of original data (\eqn{y\in(0,1)} and \eqn{y^{*}\in (-\infty,\infty)}).}

\item{mu}{mu parameter.}

\item{phi}{phi parameter.}

\item{log}{a logical value. If TRUE return the log of density function.}

\item{n}{number of observations.}
}
\value{
Return the value of density function or a random sample.
}
\description{
Density and random generation for the exponential generalized beta (EGB) of the second type,
}
\details{
For more details see
}
