% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIC.R
\docType{data}
\name{HIC}
\alias{HIC}
\title{Health Insurance Coverage (HIC)}
\format{
A data frame with 80 rows and 3 variables:
\describe{
  \item{Urbanization}{percentage of the total population who lives in the city’s urban zone}
  \item{GDP_percapita}{per capita gross domestic product}
  \item{Percent_HIC}{the health insurance coverage index}
  ...
}
}
\usage{
data("HIC",package = "robustbetareg")
}
\description{
A dataset containing the proportion of the population from several counties of the state of São Paulo in Brazil who have health insurance.
}
\keyword{datasets}
