% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show-Result.R
\name{plot.robustbetareg}
\alias{plot.robustbetareg}
\title{Interactive plots for diagnostic of robust betareg models}
\usage{
\method{plot}{robustbetareg}(object, ask = TRUE, ...)
}
\arguments{
\item{object}{fitted model object of class \code{robustbetareg}.}

\item{ask}{logical. If TRUE the user is asked before each plot.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Several types of standard diagnostic plots can be produced interactively, involving various kinds of residuals, influence measures, weights etc.
}
\examples{
\dontrun{
data("HIC", package="robustbetareg")
fit=robustbetareg(Percent_HIC~Urbanization+GDP_percapita|GDP_percapita,data=HIC,alpha=0.06)
plot(fit)}

}
