% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustbetareg.R
\name{plotenvelope}
\alias{plotenvelope}
\title{Normal Probability Plots with Simulated Envelope of Residuals for robustbetareg Objects}
\usage{
plotenvelope(
  object,
  type = c("sweighted2", "pearson", "weighted", "sweighted", "sweighted.gamma",
    "sweighted2.gamma", "combined", "combined.projection", "sweighted3"),
  conf = 0.95,
  n.sim = 100,
  PrgBar = T,
  control = robustbetareg.control(...),
  ...
)
}
\arguments{
\item{object}{Fitted model object of class \code{robustbetareg}.}

\item{type}{character indicating the type of residuals to be used, see
\code{\link{residuals.robustbetareg}}. Default is \code{type = "sweighted2"}.}

\item{conf}{numeric specifying the confidence level of the simulated
envelopes. Default is \code{conf = 0.95}.}

\item{n.sim}{a positive integer representing the number of iterations
to calculate the simulated envelopes. Default is \code{n.sim=100}.}

\item{PrgBar}{logical. If \code{PrgBar = TRUE} the progress bar will be shown
in the console.}

\item{control}{a list of control arguments specified via
\code{\link[robustbetareg:robustbetareg.control]{robustbetareg.control}}.}

\item{...}{arguments passed to \code{\link{plot}}.}
}
\value{
\code{plotenvelope} returns normal probability plot with simulated
     envelopes for the residuals.
}
\description{
\code{plotenvelope} is used to display normal probability plots with simulated
envelope of residuals for the power logit models. Currently, nine types of
residuals are supported: sweighted2, pearson, weighted, sweighted,
sweighted.gamma, sweighted2.gamma, combined, combined.projection, and
sweighted3 residuals.
}
\details{
The \code{plotenvelope} creates normal probability plots with simulated
     envelope (see Atkinson (1985) for details). Under the correct model,
     approximately 100*conf of the residuals are expected to be inside the
     envelope.
}
\examples{
\dontrun{
get(data("HIC", package = "robustbetareg"))
hic <- robustbetareg(Percent_HIC ~ Urbanization + GDP_percapita | GDP_percapita,
data = HIC, alpha = 0.06)
plotenvelope(hic, n.sim = 500)

get(data("RiskManagerCost", package = "robustbetareg"))
rmc <- robustbetareg(FIRMCOST ~ INDCOST + SIZELOG | INDCOST + SIZELOG, data = RiskManagerCost)
plotenvelope(rmc, conf = 0.90)}

}
\references{
Maluf, Y. S., Ferrari, S. L. P., and Queiroz, F. F. (2022). Robust
    beta regression through the logit transformation. \emph{arXiv}:2209.11315.\cr \cr
    Atkinson, A. C. (1985) Plots, transformations and regression: an
    introduction to graphical methods of diagnostic regression analysis.
    \emph{Oxford Science Publications}, Oxford.
}
\seealso{
\code{\link[robustbetareg:robustbetareg]{robustbetareg}}, \code{\link[robustbetareg:robustbetareg.control]{robustbetareg.control}}, \code{\link[robustbetareg:residuals]{residuals.robustbetareg}}
}
