% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(
  object,
  newdata = NULL,
  type = c("response", "link", "precision", "variance", "quantile"),
  at = 0.5
)
}
\arguments{
\item{object}{Fitted model object of class "LSMLE" or "LMDPDE" (see \code{\link[robustbetareg:robustbetareg]{robustbetareg}}).}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the original observations are used.}

\item{type}{character indicating type of predictions: fitted means of response ("response"), corresponding linear predictor ("link"), fitted precision parameter phi ("precision"),
fitted variances of response ("variance"), or fitted quantile(s) of the response distribution ("quantile").}

\item{at}{numeric vector indicating the level(s) at which quantiles should be predicted (only if type = "quantile"), defaulting to the median at = 0.5.}
}
\value{
Return a vector of predicted values.
}
\description{
Extract various types of predictions from beta regression models: either on the scale of responses in (0, 1) or the scale of the linear predictor, from LSMLE and LMDPDE objects.
}
\examples{
fit=robustbetareg(I(food/income)~income+persons|1,data=FoodExpenditure,alpha=0.08)
plotenvelope(fit,n.sim=100)

}
