% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main-Functions.R
\name{robustbetareg.control}
\alias{robustbetareg.control}
\title{Control Parameter for Robust Beta Regression}
\usage{
robustbetareg.control(
  object,
  start = NULL,
  alpha.optimal = TRUE,
  tolerance = 0.001,
  maxit = 250,
  L = 0.02,
  M = 3,
  ...
)
}
\arguments{
\item{object}{fitted model object of class "robustbetareg".}

\item{start}{a numeric vector with an initial guess of the root of estimation equation.}

\item{alpha.optimal}{a logical value. If TRUE the tuning parameter should be selected automatic.}

\item{tolerance}{numeric tolerance for convergence.}

\item{maxit}{integer specifying the maxit argument of iterations used by the Newton-Raphson algorithm.}

\item{L}{a parameter of auto selecting algorithm of tuning parameter (default L=0.02).}

\item{M}{a integer parameter value of auto selecting algorithm of tuning parameter (default M=3).}

\item{...}{currently not used.}
}
\value{
A list with the arguments specified.
}
\description{
Various parameters that control fitting of robust beta regression models using \code{\link[=robustbetareg]{robustbetareg.}}
}
\details{
For more details see:...
}
