% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-LMDPDE.R, R/Functions-LSMLE.R,
%   R/Functions-MDPDE.R, R/Functions-SMLE.R, R/Main-Functions.R
\name{LMDPDE.fit}
\alias{LMDPDE.fit}
\alias{LSMLE.fit}
\alias{MDPDE.fit}
\alias{SMLE.fit}
\alias{robustbetareg}
\title{Robust Beta Regression}
\usage{
LMDPDE.fit(y,x,z, alpha=NULL,link="logit",
link.phi="log",control=robustbetareg.control(...),...)

LSMLE.fit(y,x,z, alpha=NULL,link="logit",
link.phi="log",control=robustbetareg.control(...),...)

MDPDE.fit(y,x,z, alpha=NULL,link="logit",
link.phi="log",control=robustbetareg.control(...),...)

SMLE.fit(y,x,z, alpha=NULL,link="logit",
link.phi="log",control=robustbetareg.control(...),...)

robustbetareg(formula, data, alpha, type = c("LSMLE", "LMDPDE", "SMLE", "MDPDE"),
link=c("logit", "probit", "cloglog", "cauchit", "loglog"),
link.phi=NULL, control=robustbetareg.control(...), model = TRUE, ...)
}
\arguments{
\item{y, x, z}{\code{y} should be a numeric response vector (\eqn{y\in(0,1)}), \code{x} should be a numeric regresor matrix and \code{z} should be
a regressor matrix for the precision model, where there is the intercept only.}

\item{alpha}{the tuning value within [0,1), for robust estimation. When alpha is equal to zero (\eqn{\alpha}=0) is equivalent to MLE. If this argument is supressed, the tuning parameter should be selected automaticaly through the data-driven algorithm proposed by Ribeiro and Ferrari (2022).}

\item{link}{character specification of the link function in the mean model (mu). Currently, "logit", "probit", "cloglog", "cauchit", "loglog" are supported}

\item{link.phi}{character specification of the link function in the precision model (phi). Currently, "identity", "log", "sqrt" are supported. The default is "log" unless formula is of type y ~ x where the default is "identity"}

\item{control}{a list of control arguments specified via \code{\link[=robustbetareg.control]{robustbetareg.control}}.}

\item{...}{argument to be passed to \code{\link[=robustbetareg.control]{robustbetareg.control}}.}

\item{formula}{symbolic description of the model (of type y ~ x or y ~ x | z).}

\item{data}{arguments controlling formula.}

\item{type}{character specification of the type of estimator. Currently, LSMLE (default), LMDPDE, SMLE and MDPDE.}

\item{model}{logicals for \code{robustbetareg}. If TRUE the corresponding components of the fit (model frame, response, model matrix) are returned. For LSMLE.fit, LMDPDE.fit, SMLE.fit and MDPDE.fit \code{y} must be a numeric response vector within (0,1).}
}
\value{
robustbetareg returns an object of class "\code{robustbetareg}" with a list of the following components:\tabular{ll}{
   \code{coefficients} \tab A numeric vector of parameter estimates, \cr
   \tab \cr
   \code{vcov} \tab the covariance matrix of all parameters in the model, \cr
   \tab \cr
   \code{converged} \tab  logical indicating successful convergence of \code{optim} call, \cr
   \tab \cr
   \code{fitted.values} \tab the vector of predicted values, \cr
   \tab \cr
   \code{start} \tab the starting values for the parameters estimator, \cr
   \tab \cr
   \code{weights} \tab the weights generated by robust estimator for each obs, \cr
   \tab \cr
   \code{Tuning} \tab the selected tuning parameter, \cr
   \tab \cr
   \code{residuals} \tab a vector of standardized weighted residual 2 (see Espinheira et al (2008)), \cr
   \tab \cr
   \code{n} \tab number of observations, \cr
   \tab \cr
   \code{link} \tab mean link function applied, \cr
   \tab \cr
   \code{link.phi} \tab precision link function applied, \cr
   \tab \cr
   \code{Optimal.Tuning} \tab logical indicating whether the auto selecting tuning algorithm was selected, \cr
   \tab \cr
   \code{pseudo.r.squared} \tab pseudo R-squared value (squared correlation of linear predictor and link-transformed response), \cr
   \tab \cr
   \code{control} \tab the control arguments passed to auto selecting tuning algorithm and \code{optim} call, \cr
   \tab \cr
   \code{std.error} \tab the standard error of all parameters, \cr
   \tab \cr
   \code{method} \tab estimator type, \cr
   \tab \cr
   \code{call} \tab the original function call, \cr
   \tab \cr
   \code{formula} \tab the original formula, \cr
   \tab \cr
   \code{model} \tab the full model frame, \cr
   \tab \cr
   \code{terms} \tab a list with elemetns "\code{mean}", "\code{precision}" and "\code{full}" including the terms  \cr
   \tab objects for the respective models,  \cr
   \tab \cr
   \code{y} \tab the response proportion vector, \cr
   \tab \cr
   \code{data} \tab the dataset used. \cr
}
}
\description{
Fit robust beta regression models for rates and proportions via LSMLE, LMDPDE, SMLE and MDPDE using a parametrization with mean (depending through a link function on the covariates) and precision parameter (called phi).
}
\examples{
\dontrun{
#Table 1
data("HIC", package = "robustbetareg")
hic <- robustbetareg(Percent_HIC ~ Urbanization + GDP_percapita | 1, data = HIC, type = "LMDPDE")
summary(hic)

#Table 2
data("RiskManagerCost", package = "robustbetareg")
rmc <- robustbetareg(FIRMCOST ~ INDCOST + SIZELOG | INDCOST + SIZELOG, data = RiskManagerCost,
alpha = 0.04)
summary(rmc)}

}
\references{
\href{https://www.tandfonline.com/doi/abs/10.1080/02664760701834931}{Espinheira, P.L., Ferrari, S.L.P., and Cribari-Neto, F. (2008). On Beta Regression Residuals. Journal of Applied Statistics, 35(4), 407–419.}

\href{https://doi.org/10.1177/0962280217738142}{Ghosh, A. Robust inference under the beta regression model with application to health care studies. Statistical Methods in Medical Research, 28:271-888 (2019).}

\href{https://doi.org/10.48550/arXiv.2209.11315}{Maluf, Y. S., Ferrari, S. L., & Queiroz, F. F. (2022). Robust beta regression through the logit transformation. arXiv}

\href{https://doi.org/10.1007/s00362-022-01320-0}{Ribeiro, K. A. T. Ferrari, S. L. P. Robust estimation in beta regression via maximum Lq-likelihood. Statistical Papers (2022).}
}
\seealso{
\code{\link[robustbetareg:robustbetareg.control]{robustbetareg.control}}
}
